TensorVis/VTK: Moved VTK Output Logger to the VTK module
