+++ misc/vtk/depends.cmake
@@ -1,8 +1,6 @@
 #--------------------------------------------------------------------
 # Dependencies for current module
 # List modules on the format "Inviwo<ModuleName>Module"
-set(dependencies
-	InviwoEigenUtilsModule
-)
+set(dependencies)
 
 set(protected ON)
