+++ tensorvis/tensorvisio/src/tensorvisiomodule.cpp
@@ -41,12 +41,11 @@
 #include <modules/tensorvisio/processors/vtkunstructuredgridtorectilineargrid.h>
 #include <modules/tensorvisio/processors/flowguifilereader.h>
 #include <modules/tensorvisio/processors/vtkwriter.h>
-#include <modules/tensorvisio/util/vtkoutputlogger.h>
 
 namespace inviwo {
 
 TensorVisIOModule::TensorVisIOModule(InviwoApplication* app)
-    : InviwoModule{app, "TensorVisIO"}, vtkoutput_{std::make_unique<VtkOutputLogger>()} {
+    : InviwoModule{app, "TensorVisIO"}{
 
     registerProcessor<AmiraTensorReader>();
     registerProcessor<NRRDReader>();
